modified:   ZXCPMsrc/cpm_system.c 	modified:   ZXCPMsrc/kbd_driver.c 	modified:   ZXCPMsrc/z80.c 	modified:   ZXCPMsrc/z80.h 	modified:   ZXCPMsrc/z80_ops.c 	modified:   ZXCPMsrc/zx_display.c 	modified:   ZXCPMsrc/zx_display.h 	modified:   ZXCPMsrc/zx_system.c

+++ ZXCPMsrc/zx_display.h
@@ -24,7 +24,7 @@
 #define ZX_NEWLINE_RESET {zx_newline_flag = 0;}
 
 extern uint32_t zxlnum;
-extern uint8_t zx_newline_flag;
+extern uint32_t zx_newline_flag;
 extern uint32_t zx_border_color;
 
 void zxdisp_init();
